issues/17 - Added JavaDoc to ValidatorResource

+++ store-service/src/main/java/com/stormeye/event/repository/ValidatorRewardRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 /**
@@ -23,5 +24,5 @@ public interface ValidatorRewardRepository extends PagingAndSortingRepository<Va
     Page<ValidatorReward> findByPublicKey(final PublicKey publicKey, final Pageable request);
 
     @Query("SELECT sum(e.amount) from ValidatorReward e  WHERE e.publicKey = :publicKey")
-    long getTotalRewards(final PublicKey publicKey);
+    BigInteger getTotalRewards(final PublicKey publicKey);
 }
